Reorganise source, fill in missing columns when importing CSV
Reorganise the source code in preparation for adding database access
functions.

When importing a CSV file with a missing column name, give it a
generated name 'X?', where '?' is a monotonically increasing number
for the Lisp-Stat session.  This is similar to the behaviour to the
tidyverse read functions.

Rename parameters to csv-to-data-frame to quiet SBCL warnings about
looking like global variables.